chore: Testcontainers 환경 추상화 및 분리
