+++ src/test/java/com/team573/gongguri/global/CITests.kt
@@ -1,10 +1,9 @@
 package com.team573.gongguri.global
 
-import com.team573.gongguri.GongguriApplicationTests.Companion.mongoContainer
-import com.team573.gongguri.GongguriApplicationTests.Companion.mysqlContainer
+import com.team573.gongguri.integration.AbstractIntegrationTest
 import org.junit.jupiter.api.Test
 
-class CITests {
+class CITests : AbstractIntegrationTest(){
     @Test
     fun `컨테이너 정상 기동 확인`() {
         println("MySQL running at: ${mysqlContainer.jdbcUrl}")
