+++ src/test/java/com/team573/gongguri/integration/AbstractIntegrationTest.kt
@@ -1,0 +1,35 @@
+package com.team573.gongguri.integration
+
+import org.testcontainers.containers.MongoDBContainer
+import org.testcontainers.containers.MySQLContainer
+import org.testcontainers.junit.jupiter.Container
+import org.testcontainers.junit.jupiter.Testcontainers
+
+@Testcontainers
+abstract class AbstractIntegrationTest {
+    companion object {
+        @Container
+        val mysqlContainer = MySQLContainer<Nothing>("mysql:8.0.33").apply {
+            withDatabaseName("testdb")
+            withUsername("testuser")
+            withPassword("testpass")
+            withExposedPorts(3306)
+            start()
+            System.setProperty("spring.datasource.url", jdbcUrl)
+            System.setProperty("spring.datasource.username", username)
+            System.setProperty("spring.datasource.password", password)
+        }
+
+        @Container
+        val mongoContainer = MongoDBContainer("mongo:6.0.6").apply {
+            start()
+            val containerIpAddress = this.host
+            val port = this.getMappedPort(27017)
+
+            System.setProperty(
+                "spring.data.mongodb.uri",
+                "mongodb://$containerIpAddress:$port/testdb"
+            )
+        }
+    }
+}
